+++ crates/smtp/src/outbound/dane/verify.rs
@@ -91,7 +91,7 @@ impl TlsaVerify for Tlsa {
                 Hostname = hostname.to_string(),
             );
 
-            Err(Status::PermanentFailure(ErrorDetails {
+            Err(Status::TemporaryFailure(ErrorDetails {
                 entity: hostname.into(),
                 details: Error::DaneError("No matching certificates found in TLSA records".into()),
             }))